Atualizando sistema de login, adicionando sistema de mapa e entre outros, melhora na segurança, em fase de testes

+++ front-end/front-end-adm-react/src/components/Logs.jsx
@@ -1,6 +1,7 @@
 import { useEffect, useState } from "react";
-import { RefreshCcw } from "lucide-react";
+import { RefreshCcw, Printer } from "lucide-react";
 import Swal from "sweetalert2";
+import { DescricaoItem } from "./DescricaoItem";
 
 function formatarData(data) {
   const dataObj = new Date(data);
@@ -17,34 +18,53 @@ export function Logs() {
   const [log, setLog] = useState([]);
   const [reload, setReload] = useState(false);
 
-  function removerLog(id) {
-    Swal.fire({
-      title: "Você tem certeza?",
-      text: "Essa ação não pode ser desfeita!",
-      icon: "warning",
-      showCancelButton: true,
-      confirmButtonColor: "#d33",
-      cancelButtonColor: "#3085d6",
-      confirmButtonText: "Sim, excluir!",
-      cancelButtonText: "Cancelar",
-    }).then((result) => {
-      if (result.isConfirmed) {
-        setReload(true);
-        fetch(`https://findit-08qb.onrender.com/logs/excluir/${id}`, {
-          method: "DELETE",
-        })
-          .then(() => {
-            Swal.fire(
-              "Excluído!",
-              "O log foi removido com sucesso.",
-              "success"
-            );
-          })
-          .catch(() => {
-            Swal.fire("Erro!", "Falha ao excluir o log.", "error");
-          });
+  async function removerLog(id) {
+    const responseToken = await fetch(
+      "https://findit-08qb.onrender.com/auth-enter",
+      {
+        method: "GET",
+        credentials: "include",
       }
-    });
+    );
+
+    const dataToken = await responseToken.json();
+
+    if (dataToken.code.cargoId == 1) {
+      Swal.fire({
+        title: "Você tem certeza?",
+        text: "Essa ação não pode ser desfeita!",
+        icon: "warning",
+        showCancelButton: true,
+        confirmButtonColor: "#d33",
+        cancelButtonColor: "#3085d6",
+        confirmButtonText: "Sim, excluir!",
+        cancelButtonText: "Cancelar",
+      }).then((result) => {
+        if (result.isConfirmed) {
+          setReload(true);
+          fetch(`https://findit-08qb.onrender.com/logs/excluir/${id}`, {
+            method: "DELETE",
+          })
+            .then(() => {
+              Swal.fire(
+                "Excluído!",
+                "O log foi removido com sucesso.",
+                "success"
+              );
+            })
+            .catch(() => {
+              Swal.fire("Erro!", "Falha ao excluir o log.", "error");
+            });
+        }
+      });
+    } else {
+      Swal.fire({
+        icon: "warning",
+        title: "Atenção",
+        text: "Este usuário não possui permissão para acessar esta área. Por favor, entre em contato com o administrador.",
+        confirmButtonText: "Entendi",
+      });
+    }
   }
 
   useEffect(() => {
@@ -70,8 +90,20 @@ export function Logs() {
     }, 2000);
   }, [reload, setReload]);
 
+  const imprimirTabela = () => {
+    const printWindow = window.open("", "_blank", "width=800,height=600");
+    const content = document.getElementById("table-to-print").outerHTML;
+    printWindow.document.write(
+      "<html><head><title>Imprimir Tabela</title></head><body>"
+    );
+    printWindow.document.write(content);
+    printWindow.document.write("</body></html>");
+    printWindow.document.close();
+    printWindow.print();
+  };
+
   return (
-    <div className="w-full flex flex-col items-center overflow-y-auto h-[calc(100vh-8vh)]  px-6 py-8">
+    <div className="w-full flex flex-col items-center overflow-auto h-[calc(100vh-8vh)]  px-6 py-8">
       <div className="w-[95%]  border-b pb-4 border-gray-300 flex justify-between items-center  px-6">
         <h1 className="text-2xl font-semibold text-gray-800">
           Logs dos objetos
@@ -88,6 +120,13 @@ export function Logs() {
             }`}
           />
           <p className="font-semibold text-gray-800">Recarregar</p>
+
+          <button
+            onClick={imprimirTabela}
+            className="bg-blue-600 text-white px-1 py-1 rounded hover:bg-blue-700 translate-x-3"
+          >
+            <Printer />
+          </button>
         </div>
       </div>
 
@@ -96,37 +135,40 @@ export function Logs() {
           Nenhum Objeto Encontrado.
         </p>
       ) : (
-        <table className="w-[80%] border-collapse mt-2 font-sans">
+        <table
+          className="w-[80%] border-collapse mt-2 font-sans"
+          id="table-to-print"
+        >
           <thead>
             <tr className="bg-[#355315] text-white uppercase">
-              <th className="border border-gray-300 px-2 py-1 text-center">
-                ID Log
+              <th className="border border-gray-300 px-1 py-1 text-center text-sm">
+                Id Log
               </th>
-              <th className="border border-gray-300 px-2 py-1 text-center">
+              <th className="border border-gray-300 px-1 py-1 text-center text-sm">
                 Nome Item
               </th>
-              <th className="border border-gray-300 px-2 py-1 text-center">
+              <th className="border border-gray-300 px-1 py-1 text-center text-sm">
                 Data Adicionado
               </th>
-              <th className="border border-gray-300 px-2 py-1 text-center">
+              <th className="border border-gray-300 px-1 py-1 text-center text-sm">
                 Data Movimentação
               </th>
-              <th className="border border-gray-300 px-2 py-1 text-center">
+              <th className="border border-gray-300 px-1 py-1 text-center text-sm">
                 Achado Em
               </th>
-              <th className="border border-gray-300 px-2 py-1 text-center">
+              <th className="border border-gray-300 px-1 py-1 text-center text-sm">
                 Campus
               </th>
-              <th className="border border-gray-300 px-2 py-1 text-center">
+              <th className="border border-gray-300 px-1 py-1 text-center text-sm">
                 Situação
               </th>
-              <th className="border border-gray-300 px-2 py-1 text-center">
+              <th className="border border-gray-300 px-1 py-1 text-center text-sm">
                 Retirado Por
               </th>
-              <th className="border border-gray-300 px-2 py-1 text-center">
+              <th className="border border-gray-300 px-1 py-1 text-center text-sm">
                 Aluno CL
               </th>
-              <th className="border border-gray-300 px-2 py-1 text-center">
+              <th className="border border-gray-300 px-1 py-1 text-center text-sm no-print">
                 Ações
               </th>
             </tr>
@@ -134,40 +176,44 @@ export function Logs() {
           <tbody>
             {log.map((item, index) => (
               <tr key={index} className="hover:bg-gray-300">
-                <td className="border border-gray-300 px-2 py-1 text-center">
+                <td className="border border-gray-300 px-1 py-1 text-center text-sm">
                   {item.id_log}
                 </td>
-                <td className="border border-gray-300 px-2 py-1 text-center">
+                <td className="border border-gray-300 px-1 py-1 text-center text-sm">
                   {item.nome_item}
                 </td>
-                <td className="border border-gray-300 px-2 py-1 text-center">
+                <td className="border border-gray-300 px-1 py-1 text-center text-sm">
                   {formatarData(item.data_adicionado)}
                 </td>
-                <td className="border border-gray-300 px-2 py-1 text-center">
+                <td className="border border-gray-300 px-1 py-1 text-center text-sm">
                   {formatarData(item.data_movimentacao)}
                 </td>
-                <td className="border border-gray-300 px-2 py-1 text-center">
-                  {item.localizacao || "N/D"}
+                <td className="border border-gray-300 px-1 py-1 text-center text-sm">
+                  {item.localizacao ? (
+                    <DescricaoItem nome={item.localizacao} />
+                  ) : (
+                    "Carregando..."
+                  )}
                 </td>
-                <td className="border border-gray-300 px-2 py-1 text-center">
+                <td className="border border-gray-300 px-1 py-1 text-center text-sm">
                   {item.campus === 1
                     ? "Cotil FT"
                     : item.campus === 2
                     ? "FCA"
                     : "N/D"}
                 </td>
-                <td className="border border-gray-300 px-2 py-2 text-center">
+                <td className="border border-gray-300 px-1 py-2 text-center text-sm">
                   {item.situacao}
                 </td>
-                <td className="border border-gray-300 px-2 py-2 text-center">
+                <td className="border border-gray-300 px-1 py-2 text-center text-sm">
                   {item.retirado_por}
                 </td>
-                <td className="border border-gray-300 px-2 py-2 text-center">
+                <td className="border border-gray-300 px-1 py-2 text-center text-sm">
                   {item.clAluno}
                 </td>
-                <td className="border border-gray-300 px-2 py-2 text-center">
+                <td className="border border-gray-300 px-1 py-2 text-center text-sm no-print">
                   <button
-                    className="bg-red-600 text-white px-3 py-1 rounded hover:bg-red-700"
+                    className="bg-red-600 text-white px-2 py-1 rounded hover:bg-red-700"
                     onClick={() => removerLog(item.id_log)}
                   >
                     Remover
